moves wasm level0 functionality to seperate module
